refactor: dispatch agentphone callbacks through ccstate
