+++ turbo/apps/api/src/signals/services/agent-run-callbacks.service.ts
@@ -10,6 +10,7 @@ import { now } from "../../lib/time";
 import { db$ } from "../external/db";
 import { decryptPersistentSecretValue } from "./crypto.utils";
 import { userFeatureSwitchOverrides } from "./feature-switches.service";
+import { handleAgentPhoneInternalCallback$ } from "./internal-agentphone-run-callback.service";
 import { internalRunCallbackKindForRecord } from "./internal-run-callback";
 import { handleSlackOrgInternalCallback$ } from "./internal-slack-org-run-callback.service";
 import { handleTelegramInternalCallback$ } from "./internal-telegram-run-callback.service";
@@ -67,6 +68,19 @@ export const dispatchProgressCallbacks$ = command(
     await Promise.allSettled(
       callbacks.map(async (callback) => {
         const internalKind = internalRunCallbackKindForRecord(callback);
+        if (internalKind === "agentphone") {
+          await set(
+            handleAgentPhoneInternalCallback$,
+            {
+              callbackId: callback.id,
+              runId,
+              status: "progress",
+              payload: callback.payload,
+            },
+            signal,
+          );
+          return;
+        }
         if (internalKind === "slack:org") {
           await set(
             handleSlackOrgInternalCallback$,
